% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top_and_bottom_selection.R
\name{bottom_selection}
\alias{bottom_selection}
\title{Selecting bottom values}
\usage{
bottom_selection(table, column_name, number)
}
\arguments{
\item{table}{A data frame}

\item{column_name}{The column name you are ordering the data frame by}

\item{number}{The  number of values you want the function to return from the bottom values.}
}
\value{
An organized, miniature dataframe of bottom values from the dataframe you selected
}
\description{
Selecting the bottom values of a table ordered by the column name you enter and returns the number you specified
}
\examples{
# bottom_selection(dataframe, 'P.value', 5)

}
