% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_ontology_Gprofiler_functions.R
\name{gene_gprofile_mapping}
\alias{gene_gprofile_mapping}
\title{Mapping the gene ids to gprofiler terms}
\usage{
gene_gprofile_mapping(Table, id_col, id_to_name_map, new_col = "gene_names")
}
\arguments{
\item{Table}{Your dataframe of normalized counts, or any dataframe that has genes as the index and the gene ids}

\item{id_col}{Name of the column in df with comma-seperated gene IDs (as string)}

\item{id_to_name_map}{Named vector (names are gene IDs, values are gene names) or data frame with}

\item{new_col}{The new column name that you want for the gene names}
}
\value{
a barplot of the counts in relation to each gene ontology term.
}
\description{
Turns your dataframe into an annotated dataframe with gprofiler gene terms.
}
\examples{
# gene_gprofile_mapping(dataframe, geneID, id_to_name_mapping_list, 'gene_names')

}
