% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_ontology_Gprofiler_functions.R
\name{gprofiler}
\alias{gprofiler}
\title{gprofiler Gene ontology}
\usage{
gprofiler(eids)
}
\arguments{
\item{eids}{list of gene eids}
}
\value{
a data matrix of annotated genes using the gprofiler database
}
\description{
Makes an annotation table using gprofiler.
}
\examples{
# gprofiler(list_of_eids)

}
