% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_plotting_functions.R
\name{group_ontology_plot}
\alias{group_ontology_plot}
\title{Plotting group ontology}
\usage{
group_ontology_plot(gene_list, level = 1)
}
\arguments{
\item{gene_list}{gene list}

\item{level}{level of ontology you want to look at 1,2,3,4}
}
\value{
a plot of gene counts per ontology group
}
\description{
Function to produce ontology plots for a gene list that you specify
}
\examples{
group_ontology_plot(list_of_genes, 3)
}
