% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple_DGE_comparison.R
\name{Limma_mdge_prep}
\alias{Limma_mdge_prep}
\title{Multiple differential gene expression with Limma}
\usage{
Limma_mdge_prep(normalized_counts, metadata_table, testing_column)
}
\arguments{
\item{normalized_counts}{Dataframe of your normalized counts with sample names at the top and genes as the index.}

\item{metadata_table}{Metadata table associated with your dataframe}

\item{testing_column}{Column you are conducting the pairwise comparison in}
}
\value{
a matrix of the comparisons
}
\description{
This function runs multiple differential gene expression using the Limma package between the multiple groups you are comparing in a column
}
\examples{
# Limma_mdge_prep(normalized_counts_dataframe, metadata_table, 'Income')

}
