% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple_DGE_comparison.R
\name{Limma_mdge_results}
\alias{Limma_mdge_results}
\title{Multiple differential gene expression with Limma outputting the dataframe}
\usage{
Limma_mdge_results(normalized_counts, metadata_table, testing_column)
}
\arguments{
\item{normalized_counts}{Dataframe of your normalized counts with sample names at the top and genes as the index.}

\item{metadata_table}{Metadata table associated with your dataframe}

\item{testing_column}{Column you are conducting the pairwise comparison in}
}
\value{
A dataframe containing the differential gene expression between all groups. Returns measurements Average expression, test statistic, P. Value, adjusted P. Value, and F statistic
}
\description{
This function makes a dataframe out of the multiple differential gene expression matrix.
}
\examples{
# Limma_mdge_results(normalized_counts_dataframe, metadata_table, 'Income')

}
