% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiple_DGE_comparison.R
\name{Limma_mdge_vd_results}
\alias{Limma_mdge_vd_results}
\title{Multiple differential gene expression venn diagram}
\usage{
Limma_mdge_vd_results(normalized_counts, metadata_table, testing_column)
}
\arguments{
\item{normalized_counts}{Dataframe of your normalized counts with sample names at the top and genes as the index.}

\item{metadata_table}{Metadata table associated with your dataframe}

\item{testing_column}{Column you are conducting the pairwise comparison in}
}
\value{
a venn diagram of the matrices results
}
\description{
This function makes a venn diagram out of the multiple differential gene expression matrix.
}
\examples{
# Limma_mdge_vd_results(normalized_counts_dataframe, metadata_table, 'Income')

}
