% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata_table_creation.R
\name{new_metadata_table}
\alias{new_metadata_table}
\title{Combines two categorical columns into another metadata column}
\usage{
new_metadata_table(metadata_table, column1, column2, new_title)
}
\arguments{
\item{metadata_table}{metadata table with all the feature columns}

\item{column1}{Column 1 that you wish to combine with another column}

\item{column2}{Column 2 that you wish to combine with another column}

\item{new_title}{New title of the column you are making}
}
\value{
a dataframe of the metadata with an extra column of combined categorical values you created
}
\description{
This function combines two categorical columns into a new metadata column for pairwise comparisons
}
\examples{
#new_metadata_table(metadata, 'Ancestry', 'Sex', 'Sex_and_Ancestry')

}
