% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_plotting.R
\name{regulation_level}
\alias{regulation_level}
\title{Marking regulation}
\usage{
regulation_level(table, col1, col2, col1sig, col2sig)
}
\arguments{
\item{table}{dataframe of normalized counts}

\item{col1}{Fold change}

\item{col2}{adjusted p. value}

\item{col1sig}{significance you care about for adjusted p value}

\item{col2sig}{significance you care about for fold change}
}
\value{
a new dataframe which has a column that says upregulated or downregulated
}
\description{
Makes a new column marking upregulated or downregulated depending on p.value and log fold change
}
\examples{
regulation_level(dataframe, fold_change_col, adj_pval_col, 2.0, 0.05)

}
