% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_ontology_GOdatabase_functions.R,
%   R/Summary_stats_and_pairwise_dge.R
\name{sumstats_and_dge}
\alias{sumstats_and_dge}
\title{Calculating summary stats and pairwise gene expression}
\usage{
sumstats_and_dge(
  normalized_counts,
  metadata_table,
  testing_column,
  column1,
  column2,
  sample_char = NULL,
  sample_col_name = "Sample Name"
)

sumstats_and_dge(
  normalized_counts,
  metadata_table,
  testing_column,
  column1,
  column2,
  sample_char = NULL,
  sample_col_name = "Sample Name"
)
}
\arguments{
\item{normalized_counts}{Dataframe of your normalized counts with sample names at the top and genes as the index.}

\item{metadata_table}{Metadata table associated with your dataframe}

\item{testing_column}{Column you are conducting the pairwise comparison in}

\item{column1}{One of the categorical group names you are comparing ex: 'Male'}

\item{column2}{Another of the categorical group names you are comparing ex: 'Female'}

\item{sample_char}{The character that the sample name starts with}

\item{sample_col_name}{The column name in the metadata that has the sample names. It automatically assigns it as "Sample Name"}
}
\value{
A data matrix containing the summary stats and differential gene expression between two groups. Returns measurements means, samples, differences of each sample,  logFC, Average expression, test statistic, P. Value, adjusted P. Value, and B statistic

A data matrix containing the summary stats and differential gene expression between two groups. Returns measurements means, samples, differences of each sample,  logFC, Average expression, test statistic, P. Value, adjusted P. Value, and B statistic
}
\description{
This function calculates the summary stats and whether the two groups you are comparing have statistically significant differences in gene expression

This function calculates the summary stats and whether the two groups you are comparing have statistically significant differences in gene expression
}
\examples{
# sumstats_and_dge(normalized_counts_dataframe, metadata_table, 'Sex", 'M', 'F', 'GS', "Sample Name)

# sumstats_and_dge(normalized_counts_dataframe, metadata_table, 'Sex", 'M', 'F', 'GS', "Sample Name)

}
