% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top_and_bottom_selection.R
\name{top_selection}
\alias{top_selection}
\title{Selecting top values}
\usage{
top_selection(table, column_name, number)
}
\arguments{
\item{table}{A data frame}

\item{column_name}{The column name you are ordering the data frame by}

\item{number}{The  number of values you want the function to return from the top values.}
}
\value{
An organized, miniature dataframe of top values from the dataframe you selected
}
\description{
Selecting the top values of a table ordered by the column name you enter and returns the number you specified
}
\examples{
# top_selection(dataframe, 'P.value', 5)

}
