% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano_plotting.R
\name{volcano_regulation_expression}
\alias{volcano_regulation_expression}
\title{Volcano plotting gene expression}
\usage{
volcano_regulation_expression(
  table,
  fc_col,
  adj_pval_col,
  fc_cutoff,
  pvalue,
  gene_amount,
  title
)
}
\arguments{
\item{table}{dataframe of normalized counts}

\item{fc_col}{Fold change}

\item{adj_pval_col}{adjusted p. value}

\item{fc_cutoff}{fold change cutoff}

\item{pvalue}{p value}

\item{gene_amount}{amount of genes you want to be labeled}

\item{title}{Title of the plot you are making}
}
\value{
Volcano plot showing clearly which genes are upregulated or downregulated based on your parameters
}
\description{
Volcano plot that shows gene expression that is upregulated or downregulated between groups
}
\examples{
regulation_level(dataframe, fold_change_col, adj_pval_col, 2.0, 0.05, 30, 'African Ancestry Vs European Ancestry')

}
